SystemUI: option to toggle dimming of the notification shade (2/2)
Allow the user to enable/disable dimming of the notification shade
pulldown. Check isHighEndGfx() to respect the default setting that
the device currently supports.

Users on low-end device may want to have this feature, while users
on high-end devices may want to disable it (possible battery life
savings and what not).

Change-Id: Ib847487f8e90951a3c794137af2926da79f17696